+++ CHANGELOG.md
@@ -8,6 +8,11 @@
   - Any profile values that were previously the "raw" variant (the default) that contain template syntax (e.g. `{{user_id}}`) will now be rendered as templates. In reality this is very unlikely, so this probably isn't going to break your setup
   - If you have an existing profile value tagged with `!template` it **won't** break, but it will no longer do anything
 
+### Changed
+
+- Prevent infinite recursion in templates
+  - It now triggers a helpful error instead of a panic
+
 ## [0.17.0] - 2024-04-08
 
 ### Breaking